% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure_by_branch}
\alias{exposure_by_branch}
\title{Mutational signature exposure on the branches}
\format{
\subsection{\verb{exposure in long format}}{

A data frame with 220 rows and 3 columns,
with columns being the branch_id, the signature name, and the exposure by that signature on that branch
}
}
\source{
\url{https://github.com/YichenWang1/small_bowel/blob/main/data/mutation_matrices/sbs_mapped_to_branches.txt}
}
\usage{
exposure_by_branch
}
\description{
an SBS matrix that has been mapped to the tree nodes (branches)
from the small intestine of one patient.
The tree is stored in \code{mut_tree}
The mutation burden used to calculate this object was from \code{mut_by_branch}
}
\keyword{datasets}
