% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_contexts.R
\name{gen_contexts}
\alias{gen_contexts}
\title{generater all possible symmetric mutations given the kmer size (ie the length of the sequence)}
\usage{
gen_contexts(k)
}
\arguments{
\item{k}{an odd integer}
}
\value{
a vector of characters that represent the symmetric mutations with their corresponding kmer size
}
\description{
generater all possible symmetric mutations given the kmer size (ie the length of the sequence)
}
\examples{
gen_contexts(1)
gen_contexts(3)
gen_contexts(5)
}
