% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut.R
\name{get_context_simplified}
\alias{get_context_simplified}
\title{applied when the input has the following order
donor_id (SampleID), chromosome (Chr), position (Pos), reference_base (Ref), mutated_base (Alt)
this is in line with the simplified VCF/MAF format}
\usage{
get_context_simplified(x, seq, flank)
}
\description{
applied when the input has the following order
donor_id (SampleID), chromosome (Chr), position (Pos), reference_base (Ref), mutated_base (Alt)
this is in line with the simplified VCF/MAF format
}
