% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut.R
\name{get_context_vcf}
\alias{get_context_vcf}
\title{applied when the input has the following order
donor_id, chromosome, chrom_start, chrom_end, reference_genome_allele, mutated_from_base, mutated_to_base
this is in line with VCF/MAF format}
\usage{
get_context_vcf(x, seq, flank)
}
\description{
applied when the input has the following order
donor_id, chromosome, chrom_start, chrom_end, reference_genome_allele, mutated_from_base, mutated_to_base
this is in line with VCF/MAF format
}
