% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_on_tree.R
\name{get_signature_coordinate}
\alias{get_signature_coordinate}
\title{get signature coordinates to be plotted on the phylogenetic tree}
\usage{
get_signature_coordinate(tree, signature_burden_by_branches, bar_size = 0.1)
}
\arguments{
\item{tree}{a phylogenetic tree with each branch length being the load of mutations corresponding to that branch}

\item{signature_burden_by_branches}{a dataframe with the following required columns: 'branch' - branch/node of the tree in numeric (e.g. 1), 'signature' - the name mutational signatures  (e.g. 'SBS1'), 'exposure' - the mutation load by the corresponding signature in the corresponding branch}
}
\value{
a dataframe of signature coordinates to plot on the phylogenetic tree
}
\description{
get signature coordinates to be plotted on the phylogenetic tree
}
