% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_by_branch}
\alias{mut_by_branch}
\title{Single base substitution (SBS) mutation matrix}
\format{
\subsection{\verb{SBS mutation matrix}}{

A data frame with 33 rows and 96 columns,
with columns representing the mutations
and rows representing the tree nodes (branches)
}
}
\source{
\url{https://github.com/YichenWang1/small_bowel/blob/main/data/mutation_matrices/sbs_mapped_to_branches.txt}
}
\usage{
mut_by_branch
}
\description{
an SBS matrix that has been mapped to the tree nodes (branches)
from the small intestine of one patient.
The tree is stored in \code{mut_tree}
}
\keyword{datasets}
