% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_by_sample}
\alias{mut_by_sample}
\title{Single base substitution (SBS) mutation matrix}
\format{
\subsection{\verb{SBS mutation matrix}}{

A data frame with 21 rows and 96 columns,
with columns representing the mutations
and rows representing the samples.
This dataset is used to test
whether \code{tree_to_sample} can reconstruct
the original samples from
information that has been split to branches/nodes and its tree structure.
}
}
\source{
\url{https://github.com/YichenWang1/small_bowel/blob/main/data/mutation_matrices/sbs_small_bowel_persample.txt}
}
\usage{
mut_by_sample
}
\description{
an SBS matrix from the small intestine of one patient.
The tree is stored in \code{mut_tree}
}
\keyword{datasets}
