% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_cols.R
\name{mut_cols}
\alias{mut_cols}
\title{returns the colours based on the central mutation, colour scheme similar to COSMIC mutational signatures}
\usage{
mut_cols(kmer)
}
\arguments{
\item{kmer}{a character of mutations with their contexts}
}
\value{
the colour corresponding to the central base
}
\description{
returns the colours based on the central mutation, colour scheme similar to COSMIC mutational signatures
}
\examples{
mut_cols('A[C>T]G')
mut_cols('CT[T>G]AA')
}
