% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_tree}
\alias{mut_tree}
\title{Mutation tree}
\format{
\subsection{\verb{Newick tree}}{
}
}
\source{
\url{https://github.com/YichenWang1/small_bowel/blob/main/data/phylogenetic_trees/PD28690_snp_tree_with_branch_length.tree}
}
\usage{
mut_tree
}
\description{
a phylogenetic tree describing how samples
from one donor are related to each other.
The "tips/leaves" of the tree are the sample names
where as the nodes (branches) are estimated by MPBoot
}
\keyword{datasets}
