% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_on_tree.R
\name{plot_signature_by_branches}
\alias{plot_signature_by_branches}
\title{plot signatures on a phylogenetic tree given a matrix of mutation load}
\usage{
plot_signature_by_branches(
  tree,
  signature_burden_by_branches,
  signature_colours = NULL,
  burden = TRUE
)
}
\arguments{
\item{tree}{a phylogenetic tree with each branch length being the load of mutations corresponding to that branch}

\item{signature_burden_by_branches}{a dataframe with the following required columns: 'branch' - branch/node of the tree in numeric (e.g. 1), 'signature' - the name mutational signatures  (e.g. 'SBS1'), 'exposure' - the mutation load by the corresponding signature in the corresponding branch}

\item{signature_colours}{a vector of signature colours, named by the signatures, default to NULL (ggplot default colour pallete)}

\item{burden}{whether to display the burden axis, default to TRUE}
}
\value{
a ggplot/ggtree object of phylogenetic trees with signatures annotated
}
\description{
plot signatures on a phylogenetic tree given a matrix of mutation load
}
