% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complementary.R
\name{rv_context}
\alias{rv_context}
\title{takes a mutation context and outputs its reverse complementary counterpart}
\usage{
rv_context(context)
}
\arguments{
\item{context}{a character of mutation context}
}
\value{
a character
}
\description{
takes a mutation context and outputs its reverse complementary counterpart
}
\examples{
rv_context('A[C>G]T')
}
