% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plot.R
\name{spectra_plot}
\alias{spectra_plot}
\title{plots mutation spectra for the 3-mer}
\usage{
spectra_plot(spectra_dt, cols = NULL)
}
\arguments{
\item{spectra_dt}{spectra data, a dataframe with columns muts, kmer and value}

\item{cols}{colour vector}
}
\value{
a ggplot of spectra
}
\description{
plots mutation spectra for the 3-mer
}
