% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complementary.R
\name{strand_symmetric}
\alias{strand_symmetric}
\title{if the starting base is A or G, takes a mutation context and outputs its reverse complementary counterpart. The resulting mutations should always start with either a C or a T}
\usage{
strand_symmetric(context)
}
\arguments{
\item{context}{a character of mutation context}
}
\value{
a character
}
\description{
if the starting base is A or G, takes a mutation context and outputs its reverse complementary counterpart. The resulting mutations should always start with either a C or a T
}
\examples{
strand_symmetric('A[C>G]T')
strand_symmetric('A[A>G]T')
}
