% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_to_samples.R
\name{tree_to_samples}
\alias{tree_to_samples}
\title{recover values (mutations/exposure) that have been split into branches
back to original samples
by aggregating all values from one tip/leaf to the root
for every tip/leaf}
\usage{
tree_to_samples(branches, tree, included_columns = NULL)
}
\arguments{
\item{branches}{a dataframe of values, with a column called 'node' that corresponds to the input tree}

\item{tree}{a phylogenetic tree of Newick format, where the tip/leaf labels are the same as the original samples}

\item{included_columns}{the columns in \code{branches} to be aggregated, default to all except the column 'node'}
}
\value{
a dataframe that contains values for the original samples, these samples are the tips/leaves of the input tree
}
\description{
recover values (mutations/exposure) that have been split into branches
back to original samples
by aggregating all values from one tip/leaf to the root
for every tip/leaf
}
\examples{
# the example will use `mut_by_branch` and `mut_tree` incorporated in this package,
# original source https://github.com/YichenWang1/small_bowel
library(mutationsR)
branches = mut_by_branch
branches$node = branches$node = sapply(rownames(branches), function(x) strsplit(x, '_')[[1]][2]) # in branches, the node ID is incorporated in its rowname
tree_to_samples(branches, mut_tree)
}
