% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut.R
\name{get_mut}
\alias{get_mut}
\title{Get the central mutation from a context mutation}
\usage{
get_mut(kmer)
}
\arguments{
\item{kmer}{a character for the kmer of interest}
}
\value{
a character of mutation
}
\description{
Get the central mutation from a context mutation
}
\examples{
get_mut('A[C>T]G')

}
